// src/app/models/Ouststanding/out-standing-summary.model.ts
export class OutStandingSummary {

    Balance: number;
    BillAmount: number;
    DomainEvents: null;
    Id: number;
    PaidAmount: number;
    TotalRowCount: number;
    agentId: number;
    companyId: number;
    partyId: number;
    partyName: string;
    startDate: Date;
    endDate:Date;
}